refactor(gamemode): split modes into src/game/modes/ folder
Each game mode gets its own header (and .cpp for DataRetrievalMode).
gamemode.h now contains only the base class + GameModeId enum.
gamemode.cpp is purely the factory.
Added src/game/modes to CMake include dirs.
