+++ clients/silencer/src/game/modes/control_points_mode.h
@@ -1,0 +1,13 @@
+#ifndef CONTROL_POINTS_MODE_H
+#define CONTROL_POINTS_MODE_H
+
+#include "gamemode.h"
+
+class ControlPointsMode : public GameMode {
+public:
+	GameModeId  Id()   const override { return GAMEMODE_CONTROL_POINTS; }
+	const char* Name() const override { return "Control Points"; }
+	bool IsMatchOver(const World&) const override { return false; }
+};
+
+#endif
