+++ clients/silencer/src/game/modes/sabotage_mode.h
@@ -1,0 +1,13 @@
+#ifndef SABOTAGE_MODE_H
+#define SABOTAGE_MODE_H
+
+#include "gamemode.h"
+
+class SabotageMode : public GameMode {
+public:
+	GameModeId  Id()   const override { return GAMEMODE_SABOTAGE; }
+	const char* Name() const override { return "Sabotage"; }
+	bool IsMatchOver(const World&) const override { return false; }
+};
+
+#endif
